requirements.txt
discord.py==2.3.2
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
redis==5.0.1
minio==7.2.0
python-dotenv==1.0.0
pydantic==2.5.3
aiohttp==3.9.1
python-magic==0.4.27
requests==2.31.0
pytz==2024.1
aiosqlite==0.20.0